requirements.txt
anyio==3.6.2
asgiref==3.5.2
blinker==1.5
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
dominate==2.7.0
Flask==2.2.2
Flask-Bootstrap==3.3.7.1
Flask-Mail==0.9.1
Flask-MySQLdb==1.0.1
Flask-WTF==1.0.1
greenlet==2.0.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
httpcore==0.15.0
httpx==0.23.0
hyperframe==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
markdown2==2.4.3
MarkupSafe==2.1.1
Pillow==9.2.0
pycodestyle==2.10.0
pyjsparser==2.7.1
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-toolbelt==0.10.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.12
visitor==0.1.3
Werkzeug==2.2.2
WTForms==3.0.1
gunicorn==20.1.0
mysqlclient~=2.1.1